Integrated OpenJDK Test-suite
Added testing/tests/openjdk folder which tests compliance of JikesRVM with the Java API. The test-suite has been included in the sanity test-run but can also be run individually with --test-run openjdk. The aim of these tests is to measure progress as we attempt to transition from using the GNU Classpath libraries to the OpenJDK libraries. The success rate of the tests is currently 154/248